advanced studies project: (progress) line 70-..:removed progress com
ments. line 74-76: fixed count words instead of passing a formatted text from line 121 condition with a function of count_words to pass the text. line 79-101: c
hecking palindrome. line 80-81: replacing unecessary characters to get the words with regex, in line 84: i converted the text to string and lowered the text and
 split for proper checking, then looped through each word, then converted
 the element to list again. line 92-94: getting the first and last index of the word (which was used before the indexing) to check if both start and end are equal until i realize what palindrome really is... and then line line 98 i replaced the condition with word == word[::-1] to check if every character is exactly as backwards and forwards. line 99-102: prints if palindrome or not palindrome.